requirements.txt
-e git+https://github.com/hasadna/django.git#egg=django
boto==2.6.0
django-crispy-forms==1.2.0
django-debug-toolbar==0.9.4
django-extensions==0.9
django-haystack==2.0
git+https://github.com/tgdn/django-bootstrap-pagination.git#egg=django-bootstrap-pagination
django-nose==1.2
git+https://github.com/omab/django-social-auth.git@1a72e33dd6dade630b2fd8fb9caea765985ae015#egg=django-social-auth
django-storages==1.1.5
-e git+https://github.com/hasadna/django-taggit.git#egg=django-taggit
django-taggit-autosuggest==0.2
dj-database-url==0.2.1
gunicorn==0.15.0
ipython==0.13.1
nose==1.3.0
PIL==1.1.7
pyelasticsearch==0.2
PyYAML==3.10
requests==0.14.1
South==0.8.1
unicode-slugify==0.1
Unipath==0.2.1
wsgiref==0.1.2
yolk==0.4.3
django-flatblocks==0.7.1
selenium
mock
-e hg+https://bitbucket.org/alper/django-registration@1cddccb187ed865fd28599f0de2cc063cfa8c55a#egg=django-registration
-e git+https://github.com/daonb/django-entities#egg=django-entities
django-uuidfield
django-autoslug
-e git+https://github.com/daonb/django-chosen.git#egg=django-chosen
django-modeltranslation
whoosh==2.4
django-celery==3.0.21
celery-haystack==0.7.2
django-celery-email==1.0.3
django-devserver==0.6.2
honcho==0.4.2
facepy==0.8.4
-e git+https://github.com/jezdez/django-avatar.git#egg=django-avatar
-e git://github.com/hasadna/django-activity-stream.git@c56a07921b8c796fce35bda6feab4610fdf64e5d#egg=django-activity-stream
python-memcached
